LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/moving-between-documents/ordering/delay-load-event-1-expected.txt

CONSOLE MESSAGE: SyntaxError: The string did not match the expected pattern.
CONSOLE MESSAGE: Not executing script because it moved between documents during fetching


PASS Script elements (parser-blocking) still delay the load event in the original Document after move
PASS Script elements (parser-blocking) does not delay the load event in the new Document after move
